requirements.txt
anthropic>=0.45.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
requests>=2.31.0
